修复广播记录时间基准不一致
Why: 广播生命周期使用 FreeRTOS tick 记录，而 SNTP 格式化使用 esp_timer 回推绝对时间；混用两种计时基准会使 recorded_at 与广播时间产生偏差，32 位毫秒值长期运行还会回绕。

What: 将生命周期绝对时间戳改为 64 位 esp_timer 毫秒值，新增时间基准计算核心与主机测试。

How: 保留 tick 时钟用于扫描超时与 event_uptime_s，使用同一 esp_timer 基准生成及格式化绝对时间；执行 ./tests/run_host_tests.sh 和 idf.py build 验证。

+++ main/network/time_service.h
@@ -7,5 +7,5 @@
 void time_service_init(void);
 void time_service_start_sync(void);
 bool time_service_is_synced(void);
-bool time_service_format_wall_ms(uint32_t wall_ms, char *output, size_t output_size);
+bool time_service_format_wall_ms(uint64_t wall_ms, char *output, size_t output_size);
 const char *time_service_status_text(void);
