+++ tests/time_service_stub.c
@@ -1,7 +1,7 @@
 #include "time_service.h"
 
 bool time_service_is_synced(void) { return false; }
-bool time_service_format_wall_ms(uint32_t wall_ms, char *output, size_t output_size)
+bool time_service_format_wall_ms(uint64_t wall_ms, char *output, size_t output_size)
 { (void)wall_ms; if (output_size) output[0] = '\0'; return false; }
 void time_service_init(void) {}
 void time_service_start_sync(void) {}
